修改zone号为64bit，另外将原来zone号转化为内存拷贝的形式。 性能和原来差不多，修订了重复键的问题。

+++ db/dbformat.h
@@ -17,7 +17,7 @@
 namespace leveldb {
 
 ///////////////////////////////////////
-typedef uint32_t ZoneNumber;
+typedef uint64_t ZoneNumber;
 ///////////////////////////////////////
 
 // Grouping of constants.  We may want to make some of these
